archive/agents/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
anthropic==0.43.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
